refactor: move dto into separate packages for driver-service
