+++ driver-service/src/main/java/com/software/modsen/driverservice/controller/DriverApi.java
@@ -1,8 +1,8 @@
 package com.software.modsen.driverservice.controller;
 
-import com.software.modsen.driverservice.dto.DriverListDto;
-import com.software.modsen.driverservice.dto.DriverRequest;
-import com.software.modsen.driverservice.dto.DriverResponse;
+import com.software.modsen.driverservice.dto.response.DriverListDto;
+import com.software.modsen.driverservice.dto.request.DriverRequest;
+import com.software.modsen.driverservice.dto.response.DriverResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
