+++ driver-service/src/main/java/com/software/modsen/driverservice/service/impl/DefaultDriverService.java
@@ -1,8 +1,8 @@
 package com.software.modsen.driverservice.service.impl;
 
-import com.software.modsen.driverservice.dto.DriverListDto;
-import com.software.modsen.driverservice.dto.DriverRequest;
-import com.software.modsen.driverservice.dto.DriverResponse;
+import com.software.modsen.driverservice.dto.response.DriverListDto;
+import com.software.modsen.driverservice.dto.request.DriverRequest;
+import com.software.modsen.driverservice.dto.response.DriverResponse;
 import com.software.modsen.driverservice.exception.ResourceNotFoundException;
 import com.software.modsen.driverservice.exception.InvalidResourceException;
 import com.software.modsen.driverservice.exception.DuplicateResourceException;
